pyproject.toml
[project]
name = "spectraquant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.39.4",
    "backtrader>=1.9.78.123",
    "click>=8.1.8",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pywavelets>=1.8.0",
    "scikit-learn>=1.6.1",
    "statsmodels>=0.14.4",
    "websockets>=15.0.1",
    "yfinance>=0.2.55",
]

[project.scripts]
sq = "cli:entry"
